+++ lib/src/render/filters/filters.dart
@@ -1,0 +1,4 @@
+export 'blur_filter.dart';
+export 'color_filter.dart';
+export 'dropshadow_filter.dart';
+export 'glow_filter.dart';