+++ lib/src/utils/color_utils.dart
@@ -1,0 +1,5 @@
+import 'dart:ui';
+
+const kColorTransparent = Color(0);
+const kColorBlack = Color(0xff000000);
+const kColorRed = Color(0xff00ff00);